Set default platform so M1 macs can work (#1013)
* Set default platform so M1 macs can work

* add debug statement

* make sure the .tmx file exists

* check for map file

+++ build_maps.sh
@@ -4,6 +4,12 @@ if [ -z "$1" ]
     echo "Missing map TMX file"
     exit 1
 fi
+
+if [ ! -f "tools/maps/tmx/$1" ]; then
+  echo "File $1 does not exist"
+  exit 1
+fi
+export DOCKER_DEFAULT_PLATFORM=linux/amd64
 export PROGRESS_NO_TRUNC=1
 docker kill $(docker ps -q) 2> /dev/null
 docker build --progress plain -f Dockerfile.map . -t mapbuild --build-arg "MAP_FILE=$1" || exit 1
